Get CUDA version in run time
