+++ src/TC/inc/CudaUtils.hpp
@@ -14,10 +14,14 @@
 
 #pragma once
 
+#include "LibCuda.hpp"
 #include "TC_CORE.hpp"
 #include "tc_export.h" // Generated by cmake
-#include "LibCuda.hpp"
 #include <NppCommon.hpp>
+
+#include <mutex>
+#include <utility>
+#include <vector>
 
 namespace VPF {
 /* Check CUDA driver API return code, throw std::runtime_error with description
@@ -46,17 +50,17 @@ CUcontext GetContextByStream(CUstream str);
 int GetDeviceIdByDptr(CUdeviceptr dptr);
 
 /* Get CUDA device pointer attribute;
- * May throw exception with reason in message; 
+ * May throw exception with reason in message;
  */
 CUdeviceptr GetDevicePointer(CUdeviceptr dptr);
 
 /* Get CUDA device id by given context;
- * May throw exception with reason in message; 
+ * May throw exception with reason in message;
  */
 int GetDeviceIdByContext(CUcontext ctx);
 
 /* Get CUDA device id by given stream;
- * May throw exception with reason in message; 
+ * May throw exception with reason in message;
  */
 int GetDeviceIdByStream(CUstream str);
 
@@ -96,5 +100,32 @@ public:
 
   void Wait();
   inline CUevent Get() noexcept { return m_event; }
+};
+
+class CudaResMgr {
+private:
+  CudaResMgr();
+
+public:
+  CudaResMgr(const CudaResMgr& other) = delete;
+  CudaResMgr(const CudaResMgr&& other) = delete;
+  CudaResMgr& operator=(CudaResMgr& other) = delete;
+  CudaResMgr& operator=(CudaResMgr&& other) = delete;
+
+  CUcontext GetCtx(size_t idx);
+  CUstream GetStream(size_t idx);
+
+  int GetVersion() const;
+
+  ~CudaResMgr();
+  static CudaResMgr& Instance();
+  static size_t GetNumGpus();
+
+  std::vector<std::pair<CUdevice, CUcontext>> g_Contexts;
+  std::vector<CUstream> g_Streams;
+
+  static std::mutex gInsMutex;
+  static std::mutex gCtxMutex;
+  static std::mutex gStrMutex;
 };
 } // namespace VPF
