+++ src/python_vali/src/VALI.cpp
@@ -26,111 +26,6 @@ namespace py = pybind11;
 constexpr auto TASK_EXEC_SUCCESS = TaskExecStatus::TASK_EXEC_SUCCESS;
 constexpr auto TASK_EXEC_FAIL = TaskExecStatus::TASK_EXEC_FAIL;
 
-CudaResMgr::CudaResMgr() {
-  lock_guard<mutex> lock_ctx(CudaResMgr::gInsMutex);
-
-  ThrowOnCudaError(LibCuda::cuInit(0), __LINE__);
-
-  int nGpu;
-  ThrowOnCudaError(LibCuda::cuDeviceGetCount(&nGpu), __LINE__);
-
-  for (int i = 0; i < nGpu; i++) {
-    CUdevice cuDevice = 0;
-    CUcontext cuContext = nullptr;
-    g_Contexts.push_back(make_pair(cuDevice, cuContext));
-
-    CUstream cuStream = nullptr;
-    g_Streams.push_back(cuStream);
-  }
-  return;
-}
-
-CUcontext CudaResMgr::GetCtx(size_t idx) {
-  lock_guard<mutex> lock_ctx(CudaResMgr::gCtxMutex);
-
-  if (idx >= GetNumGpus()) {
-    return nullptr;
-  }
-
-  auto& ctx = g_Contexts[idx];
-  if (!ctx.second) {
-    CUdevice cuDevice = 0;
-    ThrowOnCudaError(LibCuda::cuDeviceGet(&cuDevice, idx), __LINE__);
-    ThrowOnCudaError(LibCuda::cuDevicePrimaryCtxRetain(&ctx.second, cuDevice),
-                     __LINE__);
-  }
-
-  return g_Contexts[idx].second;
-}
-
-CUstream CudaResMgr::GetStream(size_t idx) {
-  lock_guard<mutex> lock_ctx(CudaResMgr::gStrMutex);
-
-  if (idx >= GetNumGpus()) {
-    return nullptr;
-  }
-
-  auto& str = g_Streams[idx];
-  if (!str) {
-    auto ctx = GetCtx(idx);
-    CudaCtxPush push(ctx);
-    ThrowOnCudaError(LibCuda::cuStreamCreate(&str, CU_STREAM_NON_BLOCKING),
-                     __LINE__);
-  }
-
-  return g_Streams[idx];
-}
-
-CudaResMgr::~CudaResMgr() {
-  lock_guard<mutex> ins_lock(CudaResMgr::gInsMutex);
-  lock_guard<mutex> ctx_lock(CudaResMgr::gCtxMutex);
-  lock_guard<mutex> str_lock(CudaResMgr::gStrMutex);
-
-  stringstream ss;
-  try {
-    {
-      for (auto& cuStream : g_Streams) {
-        if (cuStream) {
-          LibCuda::cuStreamDestroy(
-              cuStream); // Avoiding CUDA_ERROR_DEINITIALIZED while
-                         // destructing.
-        }
-      }
-      g_Streams.clear();
-    }
-
-    {
-      for (int i = 0; i < g_Contexts.size(); i++) {
-        if (g_Contexts[i].second) {
-          LibCuda::cuDevicePrimaryCtxRelease(
-              g_Contexts[i].first); // Avoiding CUDA_ERROR_DEINITIALIZED while
-                                    // destructing.
-        }
-      }
-      g_Contexts.clear();
-    }
-  } catch (runtime_error& e) {
-    cerr << e.what() << endl;
-  }
-
-#ifdef TRACK_TOKEN_ALLOCATIONS
-  cout << "Checking token allocation counters: ";
-  auto res = CheckAllocationCounters();
-  cout << (res ? "No leaks dectected" : "Leaks detected") << endl;
-#endif
-}
-
-CudaResMgr& CudaResMgr::Instance() {
-  static CudaResMgr instance;
-  return instance;
-}
-
-size_t CudaResMgr::GetNumGpus() { return Instance().g_Contexts.size(); }
-
-mutex CudaResMgr::gInsMutex;
-mutex CudaResMgr::gCtxMutex;
-mutex CudaResMgr::gStrMutex;
-
 auto CopyBuffer_Ctx_Str = [](shared_ptr<CudaBuffer> dst,
                              shared_ptr<CudaBuffer> src, CUstream str) {
   if (dst->GetRawMemSize() != src->GetRawMemSize()) {
